Camera: HAL: Set the DIS, Full size liveshot modes properly.
DIS is controlled by the property persist.camera.hal.dis.
Irrespective of the value of this property, if we are running
in low power mode, DIS is disabled from lower layers. This
is used only in Normal power mode.
Full size liveshot is controlled by the property
persist.camera.full.liveshot. If the user doesnt set this
property, then irrespective of the snapshot resolution
selected from the UI, we always switch to Video size liveshot.
Also, this property value is overruled in low power mode.

Change-Id: I33679e1f1f7840ec3092fbb0802abe4bcf9830d8