camera: Handle zoom event for the snapshot path .
If we zoom in and take a snapshot in ZSL
mode, the preview frame is zoomed, but the snapshot
frame is not. This is because in ZSL mode, the HAL
preview state is PREVIEW_STARTED, but we dont process
the zoom event in that state. Change the code to
process the zoom event in PREVIEW_STARTED state if
we are in ZSL mode.

CRs-Fixed: 342796
(cherry picked from commit 5ffd516c0a0a1f8d2872966baa00a14c39a0fd26)

Change-Id: I9840656281f5e08a66517850e6cac541c3c61565