Camera: Change Auto lock to Lock & Unlock in processCtrlEvent
This change is required to avoid one of the race condition in
WNR enabled snapshot burst mode.
Deadlock is coming due to Same mLock in processCtrlEvent &
GetParameters.
Issue:
1) mlock at services has been acquired by getparameters and
datacallback is waiting for this mlcok
2) mlock at HAL is acquired by waveletdenoise and waveletdenoise
is waiting for mstopcallbacklock in notifyWDenoiseEvent
But that mstopcallbacklock is already acquired by receivecomplete
and this function is waiting for datacallback to finicsh and
datacallback is waiting for mlock acquired for getpramaters at services
Fix:
Release mLock before calling waveletdenoise

Change-Id: Id58916374d9c21087d022a073f6ea043405e953c
CRs-fixed: 468646