Camera: Making LINK_morpho_DNR_ProcessFrame Pointer to NULL
Since this is LG specific code, not able to open the LG noise
reduction .so file and so LINK_morpho_DNR_ProcessFrame pointer
is unintialized. Sometimes it is NULL and sometimes it is having
garbage value depending on the heap. This change will intialize
the pointer to NULL if there is failure in opening the .so file

Change-Id: Iebb98764cb7c09c643af81e4fe7b624986410eee
CRs-fixed: 468628