Camera: Rectified 1080p picture size dimensions.
One of the picture size in the default picture sizes array had wrong
dimension 1920x1088(instead of 1920x1080). While recording 1080p video
HAL is validating the live snapshot picture size. Since the dimension
was wrong in the available picture sizes array, it was crashing. So
updated the dimension to correct value.

Change-Id: I518e6e8ca1d8f73fd54a173e7d6cbfd937bd180a
CRs-Fixed: 523381