camera : Fix race condition in wait-signal API
Present wait-signal API implementation in HAL uses a
global strcture m_apiResult. Because of concurrent
modification of this structure from different threads,
there is a possibility of race condition where some
APIs might be waiting for ever. This change removes
resetting (memset to zero) of this global variable
to allow multiple threads to wait at the same time.

Change-Id: I8c321e39d1adf32db68534584a987bd36d9677c8