QCamera2: Properly set burst number to 1 for AOST
- Burst number for AOST need to be always set to 1,
even if previous key value is already set. The reason
is that AOST is enabled a bit later and in the beginning
burst key is already set to the property from persist
- For disabling burst number the menu setting need to be
checked directly without taking into account current zoom.
The reason is that zoom is always 1x during preview start.
- Burst setting needs to be updated after AOST features
in QCameraParameters

Change-Id: I8d0e74cbb109a23e6a1356ec955f38c4b039cea5