QCamera2: HAL: Fix the focus mode issue
Before this fix, HAL is setting the focus
mode to Auto for non Auto scene mode and
it is causing issue in some 3rd party apk's
when scene mode is changed from non auto to
Auto scene mode.
So removing the corresponding logic and
setting whatever app is calling the focus mode.

Change-Id: I51944ee1d3ab00821ab49521cb50fed57df6174f