QCamera2: Reorder and validate focus and focus_move event notifications
Focus move callbacks are not required when auto focus is triggered
in CAF mode. Also, reorder the focus move and focus done callbacks
so that even if focus state gets overriden at framework layer
(because status is stored in a single variable for both AF move and
AF done events) final status sent to application is that of AF done.

Change-Id: I24396d9b595b7f398787635535b609a06f061cab