QCamera2: Fix live snapshot thumbnail strech issue
In case of 90 or 270 degree rotation in live
snapshot, if reprocess is not enabled, HAL is
both setting the thumbnail rotation parameter
and swapping the dimensions of thumbnail which
is causing the strech effect in thumbnail.

We need to swap dimensions of thumbnail only
when source of thumbnail is main image and
rotation is already done on main image in
reprocess and of course if the rotation angle
is 90 or 270. Should not set thumbnail rotation
parameter in this case.

CRs-fixed: 751529
Change-Id: I6c67a35b0eceb7be6d7feb9725cb901e3f85b29d