QCamera2: Fix postview buffer size inconsistency
While mapping buffers to back end we need to
give same size all the times for a specific FD.
Otherwise, second map fails.

When buffers are allocated, we map those to
back end with size as 'mMemInfo[index].size'.
While reprocess, when we map the same again
in doReprocessOffline(), we are using 'frame_len'
in 'frame->bufs[i]' which should be same as
above. But in this case(postview), size received
from gralloc is saved in 'mMemInfo[index].size' and
'frame_len' is coming from 'calc_offset_postview'
which are not same and causing the issue. This
change initializes 'frame_len' as well with the size
received from gralloc.

Change-Id: I70dbe832e8b391cee6dae89663d10067856a4af5