Update recording hint to backend based on user set parameter.
In startRecording, recording hint is overideen to true if it
was false and it is set to backend also. In stopPreview HAL is
updating recording hint state by reading user set parameters.
So backend still has wrong hint value set during startRecording.
Because of this backend expects preview and video dimensions to
be of same aspect ratio, otherwise we see stretch/compression
in preview. This change updates backend with user set hint value.

Change-Id: Ia368c428d9ad6d7eea3744062f6c09636aad8bac