QCamera2: Set preview window for both main and aux
In single camera mode, we always have a preview window set. Therefore,
the same logic can be used for both display and no-display modes because
eventually preview stream will be turned on.

In dual camera mode, we want to prevent the case where aux camera streams
on preview but main camera remains streammed off for preview. This causes
sync issues for backend. To resolve this, muxer will set the preview window
for both cameras. Aux will stream on at the same time as main when the
preview window isn't NULL anymore.

Change-Id: Ib1757038531ff88c55354ee00df220bf555ca472