QCamera2:HAL3: Fix CTS testMixedBurstReprocessing.
In this testcase, there is a race condition between setparam
call for resetting perf mode and setparam for doReprocess.
Due to this race condition, the param buffer is being
overwritten and backend is not receiving doReprocess command
some times.

Here, actually, there is no call to set Perf mode from
HAL. But still reset perf call is being made because
mReprocessPerfMode variable is not initialized properly.
So, reset perf is being called even without corresponding
set perf call.

So, initialize mReprocessPerfMode variable in the reprocess
channel constructor to fix this problem.

CRs-Fixed: 981152
Change-Id: I8d67ee944fd5b58851b0122f6270754d628e8240